Moved safe_format to its own helper module
