+++ app/helpers/application_helper.rb
@@ -121,10 +121,6 @@ module ApplicationHelper
     end
   end
 
-  def safe_format(content)
-    simple_format strip_tags content
-  end
-
 private
 
   def valid_user?(user)
