+++ app/helpers/text_formatting_helper.rb
@@ -1,0 +1,8 @@
+module TextFormattingHelper
+  extend ActiveSupport::Concern
+  include ActionView::Helpers::TextHelper
+
+  def safe_format(content)
+    simple_format strip_tags content
+  end
+end
