+++ spec/helpers/application_helper_spec.rb
@@ -122,18 +122,4 @@ describe ApplicationHelper, type: :helper do
       it { expect(self.page_title).to eql('Page Title | DfE School Experience') }
     end
   end
-
-  describe '#safe_format' do
-    subject { safe_format content }
-
-    context 'with new lines' do
-      let(:content) { "hello\r\n\r\nworld" }
-      it { is_expected.to eql "<p>hello</p>\n\n<p>world</p>" }
-    end
-
-    context 'with html content' do
-      let(:content) { "<strong>hello</strong> <em>world</em>" }
-      it { is_expected.to eql "<p>hello world</p>" }
-    end
-  end
 end
